feat(multiline): properly draw strings and gaps
Instead of combining the strings, this takes one string at a time,
removes all its newlines, and then draws it to the window. Then the
current position is increased by that string's height, so I don't
have weird occurrances of an empty line in the end of the text box.