feat(graphics): chart drawing better
use the actual chart height. next is to do some better dealings with the
numbers in the chart, like maybe scaling it so that the biggest point is
always 100percent or something. Or even just forcing it to always draw
a height as a percentage of the overall window height, instead of using
the value from the module as the height itself.